PSQL LargeObjectManager methods were using longs, which caused errors. Changed to ints - needs to be tested, but as these changes only affect images it does not seem serious for now.

+++ src/mwc/facebook/data/TestPGDataStore.java
@@ -28,7 +28,7 @@ public class TestPGDataStore extends PostgresDataStore {
 			LargeObjectManager manager = ((org.postgresql.PGConnection)connection).getLargeObjectAPI();
 			
 			while(result.next()) {
-				long oid = result.getLong("image");
+				int oid = result.getInt("image");
 				System.out.printf("Photo with oid (%d) deleted\n", oid);
 				manager.delete(oid);
 			}
